requirements.txt
python-telegram-bot==20.7
requests==2.31.0
beautifulsoup4==4.12.3
aria2p==0.11.3
python-dotenv==1.0.0
aiohttp==3.9.1
aiofiles==23.2.1
tqdm==4.66.2